Update project files and dependencies for FolderCleaner
This update includes changes in configuration files due to the shifting of the FolderCleaner project to a new path. Also, a line to print the name of the directory being deleted has been added in the Program.cs file for better traceability. Furthermore, additional settings and dependencies for Microsoft.NET.ILLink.Tasks have been incorporated. The version of NuGetTool has been